requirements.txt

Django==5.0.6
Pillow
django-ckeditor
